requirements.txt
notion-client==2.0.0
PyYAML==6.0.2
mkdocs==1.4.2
mkdocs-material==9.1.6